add compatibility imports to pybikes.compat
* Move compat imports to pybikes.compat
* Move old compat functions to pybikes.data
