+++ pybikes/__init__.py
@@ -13,8 +13,8 @@ from pybikes.base import BikeShareSystem, BikeShareStation
 from pybikes.utils import PyBikesScraper
 
 # Compat old methods
-from pybikes.compat import get_data, get_all_data, get_schemas
-from pybikes.compat import get_instances, get_system_cls, get_instance
-from pybikes.compat import find_system
-from pybikes.compat import getBikeShareSystem, getDataFile, getDataFiles
+from pybikes.data import get_data, get_all_data, get_schemas
+from pybikes.data import get_instances, get_system_cls, get_instance
+from pybikes.data import find_system
+from pybikes.data import getBikeShareSystem, getDataFile, getDataFiles
 ###########################################
